feat: iterator->sysmenu,sysrole

+++ backend-admin/src/main/java/com/backend/system/entity/SysMenu.java
@@ -37,7 +37,7 @@ public class SysMenu extends BaseDO {
     /**
      * 菜单类型：0-目录；1-菜单；2-按钮
      */
-    private Long type;
+    private Integer type;
 
     /**
      * 级别：0-系统级；1-顶级top菜单；2-一级菜单；3-二级菜单；4-三级菜单
